coupling tests
