+++ noir-projects/noir-protocol-circuits/src/crates/types/src/hash.nr
@@ -324,3 +324,13 @@ pub fn compute_unique_siloed_commitments<N>(
 pub fn pedersen_hash<N>(inputs: [Field; N], hash_index: u32) -> Field {
     dep::std::hash::pedersen_hash_with_separator(inputs, hash_index)
 }
+
+#[test]
+fn compute_var_args_hash() {
+    let mut input = [0; 800];
+    for i in 0..800 {
+        input[i] = i as Field;
+    }
+    let hash = hash_args(input);
+    assert(hash == 1557627899280963684159398665725097926236612957540256425197580046184563077271);
+}
